+++ src/Glutinum.Converter/Generate.fs
@@ -33,15 +33,7 @@ let generateBindingFile (filePath: string) =
 
     let readerResult = Read.readSourceFile checker sourceFile
 
-#if DEBUG
-    printfn "glueAst: %A" readerResult.GlueAST
-#endif
-
     let res = Transform.transform true readerResult.GlueAST
-
-#if DEBUG
-    printfn "fsharpAst: %A" res
-#endif
 
     let outFile =
         {