+++ tests/specs/references/typeQuery/primitive/boolean.d.ts
@@ -1,0 +1,2 @@
+export const version : boolean;
+export function isVersion(text : typeof version): void;