+++ tests/specs/references/typeQuery/primitive/null.d.ts
@@ -1,0 +1,2 @@
+export const version : null;
+export function toVersion(text : typeof version): void;