+++ tests/specs/references/typeQuery/primitive/string.d.ts
@@ -1,0 +1,2 @@
+export const version : string;
+export function toVersion(text : typeof version): void;