+++ tests/specs/references/typeQuery/primitive/undefined.d.ts
@@ -1,0 +1,2 @@
+export const version : undefined;
+export function toVersion(text : typeof version): void;